Move test for service to service.test.ts
